refactor(shim): unify config.json access via PatchConfig
